+++ packages/sdk-core/src/api-client.ts
@@ -332,6 +332,7 @@ export class APIClient extends EventEmitter<APIClientEvents> {
 				config: {},
 				createdAt: new Date(),
 			},
+			shouldStartRecording: data.shouldStartRecording,
 		};
 	}